+++ src/Testcontainers/Builders/NpipeEndpointAuthenticationProvider.cs
@@ -6,7 +6,7 @@
   using JetBrains.Annotations;
 
   /// <inheritdoc cref="IDockerRegistryAuthenticationProvider" />
-  internal sealed class NpipeEndpointAuthenticationProvider : IDockerEndpointAuthenticationProvider
+  internal sealed class NpipeEndpointAuthenticationProvider : DockerEndpointAuthenticationProvider
   {
 #pragma warning disable S1075
 
@@ -20,13 +20,13 @@
 #pragma warning restore S1075
 
     /// <inheritdoc />
-    public bool IsApplicable()
+    public override bool IsApplicable()
     {
       return RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
     }
 
     /// <inheritdoc />
-    public IDockerEndpointAuthenticationConfiguration GetAuthConfig()
+    public override IDockerEndpointAuthenticationConfiguration GetAuthConfig()
     {
       return new DockerEndpointAuthenticationConfiguration(DockerEngine);
     }