+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - 496 Support `~/.testcontainers.properties` custom configuration
 - 500 Add trace output while building or pulling a Docker image (@michal-korniak)
 - 501 Throw an exception when Docker image has not been built (@michal-korniak)
+- 509 Check if the authentication provider can establish a Docker endpoint connection
 - 510 Add `IImageFromDockerfileBuilder.WithBuildArgument` (@michal-korniak)
 - 511 Remove temp Dockerfile archive after Docker build (@michal-korniak)
 - 512 Throw ArgumentException if Docker image name contains uppercase characters (@michal-korniak)